+++ app/routes.py
@@ -51,16 +51,13 @@ def login_employees():
     loginform = LoginForm()
     # Executando a tentativa de login quando o forms for enviado
     if request.method == "POST":
-        print("HEaaaaaaLO")
         # Validando o formulário
         if loginform.validate_on_submit():
             # Buscando os resultados no database
             attempted_user = Employees.query.filter_by(Username=loginform.Username.data).first()
-            print("HEaaaaaaLO")
             # Verificando se houve uma instância e se a senha match
             if attempted_user and attempted_user.check_password(loginform.Password.data):
                 # Fazendo o login do usuário
-                print("HELO")
                 session['user_type'] = 'employee'
                 login_user(attempted_user)
                 return redirect(url_for("schedualeemployee"))
@@ -69,7 +66,85 @@ def login_employees():
 
 @app.route("/schedualeemployee", methods=["GET", "POST"])
 def schedualeemployee():
-    return render_template("schedualeemployee.html")
+    # Requisitando a data de hoje para exibir
+    today = datetime.today()
+
+    # Colocando o dia como 1, para somar e assim usar essa variável para os demais dias
+    day1 = timedelta(days=1)
+
+    week = dict()
+    # Primeiro for: Iterar sobre os 6 dias
+    for day in range(0,7):
+        # If para pular o domingo
+        if today.weekday() == 6:
+            today = today+day1
+            continue
+        # Criando a data do dia
+        # OBS: Esse today é uma variável que vai trocando os dias
+        date = (today.date())
+        week[date] = dict()
+        # Query para encontrar os agendamentos do funcionário no dia
+
+        employee_scheduale = Classes.query.filter(and_(Classes.Employees_id==current_user.Id, func.date(Classes.Date)==today.date())).all()
+
+        # Colocando os agendamentos do funcionário no dicionário
+        for hora in employee_scheduale:
+            if hora.Account_id == 2:
+                week[date][hora.Date.hour] = [True, 2]
+            else:
+                week[date][hora.Date.hour] = [True, 1]
+        # Preenchendo as demais aulas que não foram marcadas
+        for j in range(5,24):
+            # Se já tiver com algum horário preenchido, pule
+            if j in week[date]:
+                continue
+            # Horário de almoço
+            if j == 12:
+                continue
+            week[date][j] = [False, None]
+        # Aqui fazemos a troca do dia, passa de segunda para terça, terça para quarta ...
+        today = today+day1
+        # Organizando os itens para não ficarem com um display errado no html
+        week[date] = sorted(week[date].items()) 
+
+
+    if request.method == "POST":
+            print(current_user.Id)
+            dia = str(request.form["dia"])
+            hora = int(request.form["hora"])
+            delete = str(request.form["delete"])
+
+            new_day = int(dia[-2:])
+            new_month = int(dia[-5:-3])
+            new_year = int(dia[0:4])
+            if delete == "False":
+                # realizar query para colcoar o id do user = a 0
+                date_to_free = datetime(year=new_year, 
+                                        month=new_month, 
+                                        day=new_day, 
+                                        hour=hora)
+                new_free_time = Classes(Date=date_to_free, 
+                                        Employees_id=current_user.Id, 
+                                        Account_id=2)
+                db.session.add(new_free_time)
+                db.session.commit()
+                return redirect(url_for('schedualeemployee'))
+            else:
+                # Deletar o dia que pode ser free time ou algum cliente
+                # Criando a data para excluir a folga ou a classe
+                date_to_delete = datetime(year=new_year, 
+                                        month=new_month, 
+                                        day=new_day, 
+                                        hour=hora)
+                # Query para procurar a informação no DB
+                query_to_delete = Classes.query.filter(and_(Classes.Employees_id==current_user.Id, 
+                                                            Classes.Account_id==current_user.Id, 
+                                                            Classes.Date==date_to_delete)).first()
+                db.session.delete(query_to_delete)
+                db.session.commit()
+                return redirect(url_for('schedualeemployee'))
+                
+    return render_template("schedualeemployee.html", week = week)
 
 
 
@@ -232,6 +307,9 @@ def scheduale():
             # Colocando o dia como 1, para somar e assim usar essa variável para os demais dias
             day1 = timedelta(days=1)
             # Fazendo a query do dia de hoje com aquele funcionário
+            #
+            #
+            # Excluir??
             employee_scheduale = Classes.query.filter(and_(Classes.Employees_id==employee_choosed, func.date(Classes.Date)==today.date())).all()
             week = dict()
             # Primeiro for: Iterar sobre os 6 dias